feat(artist): unify library and TIDAL artist pages
Non-library TIDAL artist pages (reached via search) rendered a bare
stub: round portrait, a "TIDAL preview" badge, a flat top-tracks list,
and an uncategorized album grid, with no play/shuffle/radio, bio,
similar artists, videos, or "See all" links.

Enrich /api/tidal/artists/{id} to return the same rich payload as the
library discography route (categorized albums, top tracks, videos,
similar artists, bio, picture) via a shared build_tidal_artist_payload
helper, then render both routes through one shared ArtistDetail
component keyed on a local id or a TIDAL id. Same treatment for the
discography "See all" pages via a shared ArtistDiscographySection
component plus a /tidal/artists/[id]/discography/[section] route.

The library code path is unchanged; TIDAL mode is additive and sources
playback from the ephemeral TIDAL play helpers. Removes the now-dead
profile-catalog helpers and repoints the artist contract tests at the
shared components.
